+++ cypress.config.ts
@@ -6,6 +6,7 @@ module.exports = defineConfig({
         baseUrl: "http://localhost:8000/",
           setupNodeEvents(on, config) {
         },
+        experimentalStudio: true,
     },
     viewportWidth: 1920,
     viewportHeight: 1200,